requirements.txt
Flask==3.1.2
gunicorn==21.2.0
flask-socketio==5.6.0
eventlet==0.40.4
flask-sqlalchemy==3.1.1
python-dotenv==1.2.1
requests==2.32.5
web3==7.14.0
pycryptodome==3.19.0
# EMV & Cryptography
cryptography==41.0.7
pyopenssl==23.3.0
# Database
# Note: `sqlite3` is part of the Python standard library and must not
# be listed in requirements.txt; remove the package entry to avoid
# installation errors on deployment platforms.
# Utilities
python-dateutil==2.8.2
# Async support
celery==5.3.4
redis==5.0.1

# gRPC / protobuf
grpcio==1.58.0
grpcio-tools==1.58.0
